beginning to organise into CLI. need to steal functionality from old main.go for WriteToCSV
