+++ go.mod
@@ -6,6 +6,11 @@ require (
 	github.com/airbusgeo/godal v0.0.10
 	github.com/golang/geo v0.0.0-20230421003525-6adc56603217
 	github.com/sirupsen/logrus v1.9.3
+	github.com/spf13/cobra v1.8.0
 )
 
-require golang.org/x/sys v0.12.0 // indirect
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	golang.org/x/sys v0.12.0 // indirect
+)
