+++ rastertoS2/rastertoS2.go
@@ -1,4 +1,4 @@
-package main
+package rastertoS2
 
 import (
 	"errors"
